ai_models/requirements.txt
flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0
openai>=1.12.0
anthropic>=0.18.1
google-cloud-aiplatform>=1.42.1
together>=0.2.8
python-dotenv==1.0.0
requests==2.31.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
pandas>=2.2.0
matplotlib>=3.8.3
tqdm>=4.66.2
wikipedia-api==0.6.0 